+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The resulting `.d.ts` are **only** compatible with typescript versions `>= 3.6`.
 ### Fixed
 
 - Fixed the result of `calculateBoundingRect` sometimes having negative width or height
+- Fixed `GroupBehaviour` interface not allowing additional properties
 
 
 ## [0.5.1] - 2020-02-26