+++ CHANGELOG.md
@@ -10,6 +10,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 **Known Issue:** this version was compiled using typescript 3.7!
 The resulting `.d.ts` are **only** compatible with typescript versions `>= 3.6`. See [typescript 3.7 release notes](www.typescriptlang.org/docs/handbook/release-notes/typescript-3-7.html#class-field-mitigations) for more information.
 
+
+## [0.5.2] - 2020-03-08
+
+**Known Issue:** this version was compiled using typescript 3.7!
+The resulting `.d.ts` are **only** compatible with typescript versions `>= 3.6`. See [typescript 3.7 release notes](www.typescriptlang.org/docs/handbook/release-notes/typescript-3-7.html#class-field-mitigations) for more information.
+
 ### Added
 
 - Added `getNodeBBox` to get the bounding box of a node
@@ -260,8 +266,9 @@ First beta version.
 - Initial webcomponent
 - Documentation
 
-[Unreleased]: https://github.com/UST-MICO/grapheditor/compare/v0.5.1...HEAD
-[0.5.0]: https://github.com/UST-MICO/grapheditor/compare/v0.5.0...v0.5.1
+[Unreleased]: https://github.com/UST-MICO/grapheditor/compare/v0.5.2...HEAD
+[0.5.2]: https://github.com/UST-MICO/grapheditor/compare/v0.5.1...v0.5.2
+[0.5.1]: https://github.com/UST-MICO/grapheditor/compare/v0.5.0...v0.5.1
 [0.5.0]: https://github.com/UST-MICO/grapheditor/compare/v0.4.1...v0.5.0
 [0.4.1]: https://github.com/UST-MICO/grapheditor/compare/v0.4.0...v0.4.1
 [0.4.0]: https://github.com/UST-MICO/grapheditor/compare/v0.3.1...v0.4.0